% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ICES2ASAP.R
\name{ICES2ASAP}
\alias{ICES2ASAP}
\title{Translating ICES file format to "vanilla" ASAP input file}
\usage{
ICES2ASAP(user.wd, user.od, model.id, ices.id)
}
\arguments{
\item{user.wd}{xxx}

\item{user.od}{xxx}

\item{model.id}{= model.id
model.id = ''}

\item{ices.id}{because sometimes there is no stock id at the beginning of the file names}
}
\description{
Translating ICES file format to "vanilla" ASAP input file
}
