% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ICES2SS.R
\name{ICES2SS}
\alias{ICES2SS}
\title{Convert ICES input files to Stock Synthesis (SS) input files}
\usage{
ICES2SS(
  user.wd,
  user.od,
  ices.id = "",
  slx = 26,
  tvslx = TRUE,
  ages = c(1, 2),
  forN = 1
)
}
\arguments{
\item{user.wd}{A file path to the directory containing the ICES files.}

\item{user.od}{A file path to a directory where the resulting files will
be saved.}

\item{ices.id}{A character value to set the ices.id because, sometimes,
there is no stock id at the beginning of the file names}

\item{slx}{A numerical value specifying the age-based selectivity type that
will be used in Stock Synthesis.}

\item{tvslx}{A logical value specifying whether or not to implement
time-varying selectivity.}

\item{ages}{Two values, a min and max age used for reporting fishing
mortality estimates for each year. These values will be placed in the
Stock Synthesis starter file for F std report ages.}

\item{forN}{An integer value specifying the number of forecast years for
the projections.}
}
\description{
A series of files are used as input to ICES models.
These files are standard input files for many types of models.
This code acts as a standardized way to convert these ICES files
to files that can be used in a Stock Synthesis (SS) model.
}
\examples{
ICES2SS(user.wd = "c:/stockAssessment/MGWG/state-space/NScod", 
  user.od = "c:/stockAssessment/MGWG/state-space/NScod/SS",
  slx = 17, tvslx = FALSE, ages = c(2,4))
output <- r4ss::SS_output(getwd(), covar = FALSE)
r4ss::SS_plots(output, uncertainty = FALSE)

}
\author{
Kelli Faye Johnson
}
