% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lastcharacter.R
\name{lastcharacter}
\alias{lastcharacter}
\title{Test equality of last character to a character value}
\usage{
lastcharacter(x, y)
}
\arguments{
\item{x}{Vector of strings}

\item{y}{A single character to test for equality to}
}
\description{
Find the last character of each string and test if it matches
the character provided in the \code{y} argument.
}
